Refactor Graph code
