+++ UpdateDevices/Program.cs
@@ -24,6 +24,7 @@ namespace UpdateDevices
           ConfigureServices(services =>
           {
             services.AddSingleton<ICosmosDbService, CosmosDbService>();
+            services.AddSingleton<IGraphService, GraphService>();
           })
           .Build();
